+++ src/Identity/Logitar.Identity.Domain/Action.cs
@@ -1,7 +1,0 @@
-﻿namespace Logitar.Identity.Domain;
-
-public enum Action
-{
-  Add = 0,
-  Remove = 1
-}