+++ src/Identity/Logitar.Identity.Domain/Passwords/IPasswordStrategy.cs
@@ -1,9 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Passwords;
-
-public interface IPasswordStrategy
-{
-  string Id { get; }
-
-  Password Create(string password);
-  Password Decode(string encoded);
-}