+++ src/Identity/Logitar.Identity.Domain/Settings/IUniqueNameSettings.cs
@@ -1,6 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Settings;
-
-public interface IUniqueNameSettings
-{
-  string? AllowedCharacters { get; }
-}