+++ src/Identity/Logitar.Identity.Domain/Settings/SettingsResolver.cs
@@ -1,9 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Settings;
-
-public class SettingsResolver : ISettingsResolver
-{
-  public virtual IRoleSettings RoleSettings { get; } = new RoleSettings();
-  public virtual IUserSettings UserSettings { get; } = new UserSettings();
-
-  public virtual IPbkdf2Settings Pbkdf2Settings { get; } = new Pbkdf2Settings();
-}