+++ src/Identity/Logitar.Identity.Domain/Users/IUserManager.cs
@@ -1,7 +1,0 @@
-﻿namespace Logitar.Identity.Domain.Users;
-
-public interface IUserManager
-{
-  Task DeleteAsync(UserAggregate user, CancellationToken cancellationToken = default);
-  Task SaveAsync(UserAggregate user, CancellationToken cancellationToken = default);
-}